% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestFunctions.R
\name{AddSpatialHarvest}
\alias{AddSpatialHarvest}
\title{Add Spatial Harvest Function}
\usage{
AddSpatialHarvest(NewxmlPath, Units_path, Gaps_path, ParamFile_Suffix)

AddSpatialHarvest(NewxmlPath, Units_path, Gaps_path, ParamFile_Suffix)
}
\arguments{
\item{NewxmlPath}{string - directory where the newly created parameter files are located}

\item{Units_path}{string - directory where the spatial files for each unit is located}

\item{Gaps_path}{string - directory where the spatial files for the gap curs are located}

\item{ParamFile_Suffix}{string - what is the ending of the parameter files - represents a given parameter file update}
}
\description{
This function differs from UpdateHarvestsFn where we used brushing. This function retains
deciduous trees in the clearcuts, with each species retained in different \% in each block.

This function differs from UpdateHarvestsFn where we used brushing. This function retains
deciduous trees in the clearcuts, with each species retained in different \% in each block.
}
