% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushingFunctions.R
\name{brush}
\alias{brush}
\title{Title}
\usage{
brush(Basexml, numBrush, timeBrush, amtBrush = 98)
}
\arguments{
\item{numBrush}{a vector of the harvest numbers that correspond to brushing}

\item{timeBrush}{Which year(s) to brush}

\item{amtBrush}{\% of trees to brush}
}
\description{
Title
}
\details{
could add another parameter to define lower and upper limit of brushing
}
