% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{calc_sph_92data}
\alias{calc_sph_92data}
\title{summarize live or dead trees by unit}
\usage{
calc_sph_92data(
  diam_sizeClass = 2.5,
  dateCreek92_data,
  dateCreek_92_tallies,
  liveTrees = TRUE
)
}
\arguments{
\item{liveTrees}{}
}
\description{
summarize the stems/ha by species and diameter class (and tree class if snags) by unit in
the Date Creek Silvicultural Trial. This tree list from the 1992 pre-treatment data can be used to initiate forest
models such as SORTIE. Use
}
\author{
Kiri Daust, Erica Lilles, Jonathan VanElslander and Alana Clason worked on this function
}
