% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFunctions.R
\name{calculateSPH}
\alias{calculateSPH}
\title{Calculate stems per hectare from prism plots}
\usage{
calculateSPH(pBands, DBH)
}
\arguments{
\item{pBands}{which prism band used (numeric)}

\item{DBH}{diameter of tree}
}
\value{
The equivalent stems/ha of a given diameter
}
\description{
Calculate stems per hectare from prism plots
}
