% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_Fielderror_corrections.R
\name{corr_trees_2018}
\alias{corr_trees_2018}
\title{Internal function to correct 2018 tree data}
\usage{
corr_trees_2018(dat)
}
\arguments{
\item{dat}{}
}
\description{
Internal function to correct 2018 tree data
}
