% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_Fielderror_corrections.R
\name{corr_trees_2019}
\alias{corr_trees_2019}
\title{Internal function to correct 2019 tree data}
\usage{
corr_trees_2019(dat)
}
\arguments{
\item{dat}{}
}
\description{
Internal function to correct 2019 tree data
}
