% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFunctions.R
\name{createTreeListSortie}
\alias{createTreeListSortie}
\title{Create an initial tree list for SORTIE}
\usage{
createTreeListSortie(
  diam_sizeClass = 2.5,
  dateCreek92_data,
  dateCreek_92_tallies,
  liveTrees = TRUE
)
}
\arguments{
\item{diam_sizeClass}{what size diameter classes to use for SORTIE. Default set to 2.5. If changing to 2, this code
will be inaccurate as the lower DBH limit for the prism plots was 7.5.}

\item{dateCreek92_data}{filename (and pathway) for the 1992 Date Creek prism data}

\item{dateCreek_92_tallies}{file name (and pathway) for the 1992 Date Creek small tree fixed radius plots data}

\item{liveTrees}{Do you want live trees (TRUE = default) or snags (FALSE)}
}
\description{
This function creates an output compatible with SORTIE from dat92_sph_diamClass datatable. You can either
pass the output data table explicitly, or this function will run the dat92_sph_diamClass function if not already available

If you change how the species are spelled in a SORTIE parameter file (i.e. Western_Hemlock become westernHemlock),
you would have to change it here - the species spelling is hard coded here and MUST MATCH EXACTLY what is in the
associated parameter file
}
