% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_1992_diams}
\alias{CWD_1992_diams}
\title{Calculate mean diameters by size class for 1992 CWD by Unit}
\usage{
CWD_1992_diams(CWD_dat)
}
\arguments{
\item{CWD_dat}{}
}
\description{
Calculate mean diameters by size class for 1992 CWD by Unit
}
