% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_1993_props}
\alias{CWD_1993_props}
\title{Calculate CWD proportions after harvest from Date Creek study}
\usage{
CWD_1993_props(dat_loc)
}
\arguments{
\item{dat_loc}{}
}
\description{
Calculate CWD proportions after harvest from Date Creek study
}
