% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_2011_Vol_cal}
\alias{CWD_2011_Vol_cal}
\title{2011 CWD volume calculation}
\usage{
CWD_2011_Vol_cal(CWD_dat, Horiz_dat)
}
\arguments{
\item{CWD_dat}{}
}
\description{
2011 CWD volume calculation
}
\details{
Calculate Volume using the British Columbia Ministry of Forests and Range (2009) formula
CWD volume (m3/ha) = pi^2/8L  *  sum\link{D^2/cos (A)}
Where: 	L = length of total transect (horizontal distance (HD) in m)
HD = SD / Square root of \link{1 + (\% slope / 100)2}
D = diameter of each piece of CWD (cm)
A = tilt angle from horizontal for each piece (degrees)

There is only one horizontal distance file - works for both 2011 and 2018?
}
