% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_2018_Vol_calc}
\alias{CWD_2018_Vol_calc}
\title{2018 CWD function}
\usage{
CWD_2018_Vol_calc(CWD_dat, out_carbon_comp = FALSE)
}
\arguments{
\item{CWD_dat}{}
}
\value{
data.table with total volume of CWD by plot and Unit
}
\description{
\code{CWD_2018_Vol_calc()}
}
\details{
2018 Date Creek coarse woody SQUARE PLOTS (not transects)

Alternative volume calculation (not used - but here just in case want to play with)
Volume calculation - taper volume equation (used in Gove paper)Using the taper equation results in the most similar volume values as the volume from transects
Functions for calculating gove volume

Using conic-parabloid equation from Fraver et al. 2013
length and diameter in meters
Bole volume equations

Assign relevant data to the equation components
du <- CWD.2018$small_diam #small diameter
db <- CWD.2018$large_diam #large diameter
L  <- CWD.2018$Lngth2end #total length
il <- L/2 #intermediate log length
r  <- 3 #paraboloid

Calculate volumes of the truncated boles
taper <- taper_volume(du,db,L,r,il)
}
\author{
Ingrid Farnell
}
