% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_2019_Vol_calc}
\alias{CWD_2019_Vol_calc}
\title{2019 function:}
\usage{
CWD_2019_Vol_calc(CWD_dat, Horiz_dat, out_carbon_comp = FALSE)
}
\arguments{
\item{Horiz_dat}{}
}
\value{
data.table with total volume of CWD by plot and Unit
}
\description{
\code{CWD_2019_Vol_calc()}
}
\details{
2019 Date Creek coarse woody transects - repeat measure from 2011

Calculate Volume using the British Columbia Ministry of Forests and Range (2009) formula
CWD volume (m3/ha) = pi^2/8L  *  sum\link{D2/cos (A)}
Where: 	L = length of total transect (horizontal distance (HD) in m) HD = SD / Square root of \link{1 + (\% slope / 100)2}
slope was not measured so assume total transect length to be 90m
D = diameter of each piece of CWD (cm)
A = tilt angle from horizontal for each piece (degrees)
Because tilt angle was not measured in 2019, it was assumed to be zero, so that cos (A) = 1 for all pieces in those years.
}
\author{
Ingrid Farnell
}
