% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_vol_calc}
\alias{CWD_vol_calc}
\title{Date Creek CWD - all years}
\usage{
CWD_vol_calc(dat_loc, incl_sp_decay = FALSE)
}
\arguments{
\item{dat_loc}{what is the directory where all data is stored}
}
\description{
Date Creek CWD - all years
}
\details{
we expect the directory to contain 6 files: CWD_1992.csv, CWD_1993.csv, CWD_2011.csv, CWD_2018.csv and CWD_2019.csv
}
