% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateCreekData-package.R
\docType{package}
\name{DateCreekData-package}
\alias{DateCreekData}
\alias{DateCreekData-package}
\title{DateCreekData: Date Creek Data Processing}
\description{
Functions to prepare Date Creek data for common summaries and use with sortie and analyses including carbon accounting.
}
\author{
\strong{Maintainer}: Alana Clason \email{alana.clason@bvcentre.ca}

Authors:
\itemize{
  \item Ingrid Farnell \email{farnellingrid@gmail.com}
}

}
\keyword{internal}
