% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planting.R
\name{InitPlant}
\alias{InitPlant}
\title{Initiate Date Creek planting}
\usage{
InitPlant(
  My_newvalsPath,
  PlantDensDat,
  PlantPCDat,
  Plant_param_name,
  InitValues_csv_end,
  update_plant_pos_only
)
}
\arguments{
\item{update_plant_pos_only}{}
}
\description{
Initiate Date Creek planting
}
