% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampGridOutputs.R
\name{maskGrids}
\alias{maskGrids}
\title{Sample the Date Creek Grid outputs from SORTIE}
\usage{
maskGrids(
  Blocks = DateCreekData::Treatments$Unit,
  Units_path = Units_path,
  Gaps_path = Gaps_path,
  NoCells_ToSample = NA,
  grid_dat,
  output = "table",
  grid_to_output
)
}
\arguments{
\item{Blocks}{treatment blocks}

\item{Units_path}{where are the unit boundaries}

\item{Gaps_path}{where are the gap boundaries}

\item{NoCells_ToSample}{defaul = NA (no sampling), otherwise numeric value}

\item{grid_dat}{data.table object containing parsed grid outputs}

\item{output}{output type: "raster" or "table"}

\item{grid_to_output}{name(s) of grid(s) to output}
}
\description{
Sample the Date Creek Grid outputs from SORTIE
}
