% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialBoundaries.R
\name{modifyPartCuts}
\alias{modifyPartCuts}
\title{Modify the Date Creek treatment boundaries for heavy and light removals}
\usage{
modifyPartCuts(Gaps_path)
}
\arguments{
\item{Gaps_path}{The directory with the kmls and kmzs that define the cuts}
}
\value{
an sf object
}
\description{
Modify the Date Creek treatment boundaries for heavy and light removals
}
