% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialBoundaries.R
\name{prepSpatialBounds}
\alias{prepSpatialBounds}
\title{read and prep harvest boundaries for sortie}
\usage{
prepSpatialBounds(Units_path, Gaps_path)
}
\arguments{
\item{Gaps_path}{}
}
\description{
not sure we actually need this function - would need to return 2 objects...
}
