% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialBoundaries.R
\name{read_write_harvests_xmls}
\alias{read_write_harvests_xmls}
\title{BROKEN!!!!! write out harvest and planting for xml updates}
\usage{
read_write_harvests_xmls(Units_path, bounds, partCuts, Unit_i, PlotSize)
}
\arguments{
\item{PlotSize}{}
}
\description{
BROKEN!!!!! write out harvest and planting for xml updates
}
