% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFunctions.R
\name{ReadSpatialBounds}
\alias{ReadSpatialBounds}
\title{Read the Date Creek spatial unit boundaries}
\usage{
ReadSpatialBounds(Units_path)
}
\arguments{
\item{Units_path}{}
}
\description{
Read the Date Creek spatial unit boundaries
}
