% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_Fielderror_corrections.R
\name{remove_plots_trees}
\alias{remove_plots_trees}
\title{Remove trees or plots from data}
\usage{
remove_plots_trees(dat)
}
\arguments{
\item{dat}{}
}
\description{
Remove trees or plots from data
}
\examples{
eliminates trees or plots from data to match missing plots from other years

}
