% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFunctions.R
\name{SampleDateCreekGrids}
\alias{SampleDateCreekGrids}
\title{Sample the Date Creek Grid outputs from SORTIE}
\usage{
SampleDateCreekGrids(
  Blocks,
  UnitBounds,
  HR_gaps,
  NoCells_ToSample,
  sampl_rast = "Y"
)
}
\arguments{
\item{NoCells_ToSample}{}
}
\description{
Sample the Date Creek Grid outputs from SORTIE
}
