% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFunctions.R
\name{subplot_outputs}
\alias{subplot_outputs}
\title{subplot_outputs}
\usage{
subplot_outputs(
  out_path,
  run_name,
  Units_path,
  yrs,
  dist_edge = 20,
  num_subplots = 30,
  size_subplot = 7.98,
  plot_treatments = TRUE
)
}
\description{
subplot_outputs
}
