% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_1992}
\alias{trees_1992}
\title{Clean 1992 tree data}
\usage{
trees_1992(
  cruise_data = "./data-raw/Trees/1992data.csv",
  fixed_data = "./data-raw/Trees/1992fixed_radius_data_fromTable20_DateCkHandbook.csv",
  calc_height = TRUE
)
}
\arguments{
\item{fixed_data}{}

\item{calc_height}{can be TRUE, "cruise" or "both"}
}
\description{
cleans tree data and calculates heights (if calc_height = TRUE)
Note - there are 2 trees in 1992 that are dead (tree class 4) that are stubs with a height of 0
}
