% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_1993}
\alias{trees_1993}
\title{Clean 1993 tree data}
\usage{
trees_1993(data = "./data-raw/Trees/SS93forR.csv", calc_height = TRUE)
}
\arguments{
\item{calc_height}{}
}
\description{
Clean 1993 tree data
}
