% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFunctions.R
\name{UpdateHarvestsFn}
\alias{UpdateHarvestsFn}
\title{Update Harvest Function}
\usage{
UpdateHarvestsFn(NewxmlPath, Units_path, Gaps_path, ParamFile_Suffix)
}
\arguments{
\item{NewxmlPath}{string - directory where the newly created parameter files are located}

\item{Units_path}{string - directory where the spatial files for each unit is located}

\item{Gaps_path}{string - directory where the spatial files for the gap curs are located}

\item{ParamFile_Suffix}{string - what is the ending of the parameter files - represents a given parameter file update}
}
\description{
Formerly UpdateHarvestsFn, This function adds the spatial harvest details for each block
with brushing. This function will be removed if we no longer wish to brush and replace with retained
deciduous trees in clear cuts (addSpatialHarvest)
}
