% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialBoundaries.R
\name{read_keyhole}
\alias{read_keyhole}
\title{(helper function) Read keyhole (kml) files into sf objects}
\usage{
read_keyhole(file)
}
\arguments{
\item{file}{filename (character)}
}
\value{
an sf object
}
\description{
(helper function) Read keyhole (kml) files into sf objects
}
