% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFunctions.R
\name{ReadSpatialBounds}
\alias{ReadSpatialBounds}
\title{Import the Date Creek treatment boundaries}
\usage{
ReadSpatialBounds(Units_path)
}
\arguments{
\item{Units_path}{The directory with the shapefiles that define the edge of the boundaries}
}
\value{
an sf object with
}
\description{
Import the Date Creek treatment boundaries
}
