% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_1992}
\alias{trees_1992}
\title{Clean 1992 tree data - doesn't work as a function b/c of plyr}
\usage{
trees_1992(
  cruise_data = "./data-raw/Trees/1.1992data.csv",
  fixed_data = "./data-raw/Trees/2.1992fixed_radius_data_fromTable20_DateCkHandbook.csv",
  calc_height = TRUE
)
}
\arguments{
\item{calc_height}{description}
}
\description{
cleans tree data and calculates heights (if height = TRUE)
can't get this working because of ddply - needs to be updated to either dplyr or data.table
}
