% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_2010}
\alias{trees_2010}
\title{Title}
\usage{
trees_2010(
  lrg_trees = "./data-raw/Trees/Data Creek 2010 Data large trees.csv",
  cc_trees = "./data-raw/Trees/Trees 10cm and above in clearcut 2010.csv",
  small_trees = "./data-raw/Trees/Date Creek 2010 Trees less than 10 cm tallies.csv",
  snag_heights = "./data-raw/Trees/SnagHeights2010.csv",
  measured_heights2022 = "./data-raw/Trees/StandStructureData_Nov2022_Final.xlsx",
  calc_height = TRUE,
  use_standard = FALSE,
  id_gap_trees = TRUE,
  use_size = TRUE,
  planted_hgt = 20,
  planted_diam = 20,
  sp_plant = "Sx",
  use_cc = TRUE
)
}
\arguments{
\item{calc_height}{}
}
\description{
Title
}
\details{
cruised_hgt in the output is only relevant for dead trees (snags) - these heights are actually
from 2022 measurements.
}
