% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_201x}
\alias{trees_201x}
\title{Title}
\usage{
trees_201x(
  data_file = "./data-raw/Trees/Date Creek 2018 Data large trees_re-entered.xlsx",
  data_2018 = "DataCk re-entry 2018 largeTrees",
  data_2019 = "./data-raw/Trees/Data Creek 2019 Data large trees.csv",
  inter_trees = "./data-raw/Trees/2018-19intermediatetrees.csv",
  small_trees = "./data-raw/Trees/Small trees 2018 2019 KHP.csv",
  lrg_trees_2010 = "./data-raw/Trees/Data Creek 2010 Data large trees.csv",
  measured_heights2022 = "./data-raw/Trees/StandStructureData_Nov2022_Final.xlsx",
  snag_hgts_18 = "./data-raw/Trees/SnagHeights2018.csv",
  calc_height = TRUE
)
}
\arguments{
\item{calc_height}{}
}
\description{
Title
}
