% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_201x}
\alias{trees_201x}
\title{Title}
\usage{
trees_201x(
  data_2018 = "./data-raw/Trees/7.Data Creek 2018 Data large trees.csv",
  data_2019 = "./data-raw/Trees/8.Data Creek 2019 Data large trees.csv",
  int_trees = "./data-raw/Trees/9.2018-19intermediatetrees.csv",
  sm_trees = "./data-raw/Trees/10.Small trees 2018 2019 KHP.csv",
  calc_height = TRUE
)
}
\arguments{
\item{calc_height}{}
}
\description{
Title
}
