% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{trees_2022}
\alias{trees_2022}
\title{2022 data}
\usage{
trees_2022(
  data_file = "./data-raw/Trees/StandStructureData_Nov2022_Final.xlsx",
  large_trees = "Large",
  inter_trees = "Inter",
  small_trees = "Small",
  calc_height = TRUE
)
}
\arguments{
\item{large_trees}{name of sheet with large trees}

\item{inter_trees}{name of sheet with intermediate trees}

\item{small_trees}{name of sheet with small trees}

\item{calc_height}{will only be applied to trees with NA for height. Standard dbh-height allometry}
}
\description{
2022 data
}
\details{
the measured heights are considered good here, so we are using them unless the height is NA
}
